move onsubmit to lessen drilling
