+++ services/ui-src/src/components/reports/sections/StaticDrawerSection.test.tsx
@@ -11,7 +11,6 @@ import {
   RouterWrappedComponent,
 } from "utils/testing/setupJest";
 
-const mockOnSubmit = jest.fn();
 const mockUseNavigate = jest.fn();
 jest.mock("react-router-dom", () => ({
   useNavigate: () => mockUseNavigate,
@@ -26,7 +25,6 @@ const staticDrawerSectionComponent = (
       <StaticDrawerSection
         form={mockForm}
         drawer={mockPageJsonStaticDrawer.drawer}
-        onSubmit={mockOnSubmit}
       />
     </ReportContext.Provider>
   </RouterWrappedComponent>
