+++ services/ui-src/src/components/reports/sections/StaticPageSection.test.tsx
@@ -9,7 +9,6 @@ import {
   RouterWrappedComponent,
 } from "utils/testing/setupJest";
 
-const mockOnSubmit = jest.fn();
 const mockUseNavigate = jest.fn();
 jest.mock("react-router-dom", () => ({
   useNavigate: () => mockUseNavigate,
@@ -21,7 +20,7 @@ jest.mock("react-router-dom", () => ({
 const staticPageSectionComponent = (
   <RouterWrappedComponent>
     <ReportContext.Provider value={mockReportContext}>
-      <StaticPageSection form={mockForm} onSubmit={mockOnSubmit} />
+      <StaticPageSection form={mockForm} />
     </ReportContext.Provider>
   </RouterWrappedComponent>
 );
